feat: configures the flex grid
