+++ src/app/homepage/homepage.component.html
@@ -1,2 +1,1 @@
-<div class="home">homepage works!</div>
 <app-list></app-list>
