+++ src/app/homepage/homepage.component.ts
@@ -1,12 +1,5 @@
-import { Component, OnInit, NgModule } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 
-export interface Tile {
-  cols: number;
-  rows: number;
-  text?: string;
-  color?: string;
-  id?: string;
-}
 @Component({
   selector: 'app-homepage',
   templateUrl: './homepage.component.html',
@@ -16,12 +9,4 @@ export interface Tile {
 export class HomepageComponent implements OnInit {
     constructor() {}
     ngOnInit() {}
-    tiles: Tile[] = [
-      {text: 'One', cols: 2, rows: 5, id: 'primaryCard'},
-      {text: 'Two', cols: 1, rows: 2, id: 'secondCard'},
-      {text: 'Three', cols: 1, rows: 2, id: 'thirdCard'},
-      {text: 'Four', cols: 1, rows: 2, id: 'fourthCard'},
-      {text: 'Five', cols: 1, rows: 2, id: 'fifthCard'},
-    ];
-
 }
